+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2023-12-06
+
+### Added
+- Added mesuring date, processing date, version, sampling rate, filename , and source to analysis JSON.
+
 
 ## [0.10.0] - 2023-11-28